backend/requirements.txt
Flask
spacy
pdfminer.six
flask-cors
python-docx